requirements.txt
streamlit
rdkit-pypi
pandas
numpy
scipy
requests
